Move FlutterMapStateContainer in to FlutterMap's file since it is just the widget state
